requirements-dev.txt
-r requirements.txt
pytest==6.2.5
requests-mock==1.10.0
